Bugsnag reporting: Move bugsnag hooks file to bugsnag config file
