+++ bin/maze-runner
@@ -8,6 +8,7 @@ require_relative '../lib/maze'
 require_relative '../lib/maze/appium_server'
 require_relative '../lib/maze/browser_stack_devices'
 require_relative '../lib/maze/browser_stack_utils'
+require_relative '../lib/maze/bugsnag_config'
 require_relative '../lib/maze/capabilities'
 require_relative '../lib/maze/compare'
 require_relative '../lib/maze/docker'
@@ -34,7 +35,6 @@ require_relative '../lib/maze/assertions/request_set_assertions'
 require_relative '../lib/maze/hooks/hooks'
 require_relative '../lib/maze/hooks/appium_hooks'
 require_relative '../lib/maze/hooks/browser_hooks'
-require_relative '../lib/maze/hooks/bugsnag_hooks'
 require_relative '../lib/maze/hooks/command_hooks'
 
 require_relative '../lib/maze/driver/appium'
